pyproject.toml
[project]
name = "owls"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyautogen>=0.10.0",
    "requests>=2.32.4",
    "serena-agent",
    "pyyaml>=6.0.0",
    "autogen>=0.9.8.post1",
    "click>=8.0.0",
    "tomli>=2.0.0; python_version<'3.11'",
]

[tool.uv.sources]
serena-agent = { git = "https://github.com/oraios/serena.git" }
